Rename pipeline_nodes to pipeline_stages
